refactor(/src): 패키지 개선
