+++ src/main/java/coupon/infrastructure/audit/JpaAuditingConfiguration.java
@@ -1,4 +1,4 @@
-package coupon.common.audit;
+package coupon.infrastructure.audit;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
