+++ src/main/java/coupon/infrastructure/datasource/DataSourceConfiguration.java
@@ -1,4 +1,4 @@
-package coupon.support.datasource;
+package coupon.infrastructure.datasource;
 
 import java.util.Map;
